Added get_root, get_cwd, and set_cwd.

+++ commands.cpp
@@ -71,8 +71,8 @@ void fn_lsr (inode_state& state, const wordvec& words){
 }
 
 void fn_make (inode_state& state, const wordvec& words){
-   DEBUGF ('c', state);
-   DEBUGF ('c', words);
+    DEBUGF ('c', state);
+    DEBUGF ('c', words);
 }
 
 void fn_mkdir (inode_state& state, const wordvec& words){
@@ -83,7 +83,7 @@ void fn_mkdir (inode_state& state, const wordvec& words){
 // Changes the prompt character to something the user sets.
 void fn_prompt (inode_state& state, const wordvec& words){
    string newPrompt = "";
-   for(int i = 1; i < words.size(); ++i) newPrompt += words.at(i);
+   for(size_t i = 1; i < words.size(); ++i) newPrompt += words.at(i);
    // If there's no trailing space, add one for looks.
    if(words.at(words.size()-1) != " ") newPrompt += ' ';
    state.setPrompt(newPrompt);
